Init frontend
