+++ frontend/app/build.gradle.kts
@@ -5,11 +5,11 @@ plugins {
 }
 
 android {
-    namespace = "com.nasti.bookingapp"
+    namespace = "com.nasti.frontend"
     compileSdk = 35
 
     defaultConfig {
-        applicationId = "com.nasti.bookingapp"
+        applicationId = "com.nasti.frontend"
         minSdk = 24
         targetSdk = 35
         versionCode = 1
@@ -27,46 +27,28 @@ android {
             )
         }
     }
-
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_11
         targetCompatibility = JavaVersion.VERSION_11
     }
-
     kotlinOptions {
         jvmTarget = "11"
     }
-
     buildFeatures {
         compose = true
     }
 }
 
 dependencies {
-    // AndroidX Core
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
-
-    // Jetpack Compose
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.ui)
     implementation(libs.androidx.ui.graphics)
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.material3)
-
-    // AndroidX AppCompat (Fixed Syntax)
-    implementation(libs.androidx.appcompat)
-
-    // Lifecycle & ViewModel
-    implementation("androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1")
-    implementation("androidx.lifecycle:lifecycle-livedata-ktx:2.5.1")
-
-    // Retrofit (Networking)
-    implementation("com.squareup.retrofit2:retrofit:2.9.0")
-    implementation("com.squareup.retrofit2:converter-gson:2.9.0")
-
-    // Testing
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
