+++ frontend/app/src/androidTest/java/com/nasti/frontend/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.nasti.bookingapp
+package com.nasti.frontend
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.nasti.bookingapp", appContext.packageName)
+        assertEquals("com.nasti.frontend", appContext.packageName)
     }
 }
