--- frontend/app/src/main/java/com/nasti/bookingapp/model/User.kt
@@ -1,7 +1,0 @@
-package com.nasti.bookingapp.model
-
-data class User(
-    val id: Int,
-    val name: String,
-    val email: String
-)
