--- frontend/app/src/main/java/com/nasti/bookingapp/repository/UserRepository.kt
@@ -1,35 +1,0 @@
-package com.nasti.bookingapp.repository
-
-/*import com.nasti.bookingapp.api.ApiService
-import com.nasti.bookingapp.model.User
-import retrofit2.Response
-
-class UserRepository {
-    private val apiService = ApiService.create()
-
-    suspend fun fetchUsers(): List<User>? {
-        val response: Response<List<User>> = apiService.getUsers()
-        return if (response.isSuccessful) response.body() else null
-    }
-}*/
-
-import android.util.Log
-import com.nasti.bookingapp.api.ApiService
-import com.nasti.bookingapp.model.User
-import retrofit2.Response
-
-class UserRepository {
-    private val apiService = ApiService.create()
-
-    suspend fun fetchUsers(): List<User>? {
-        val response: Response<List<User>> = apiService.getUsers()
-
-        if (!response.isSuccessful) {
-            Log.e("UserRepository", "Error: ${response.errorBody()?.string()}")
-            return null
-        }
-
-        Log.d("UserRepository", "Fetched users: ${response.body()}")
-        return response.body()
-    }
-}
