+++ frontend/app/src/main/java/com/nasti/frontend/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.nasti.bookingapp.ui.theme
+package com.nasti.frontend.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -34,7 +34,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun MyApplicationTheme(
+fun FrontendTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
